requirements.txt
tensorflow==2.5.0
opencv_python==4.5.2.52
streamlit==0.82.0
numpy==1.20.3
Pillow==8.2.0
